requirements.txt
torch
torchvision
scikit-learn
pillow
matplotlib
numpy
fastapi
uvicorn
python-multipart
